Add spacefm-1.0.0.
spacefm-1.0.0.ebuild is copied from spacefm-0.9.4.ebuild and is modified
to also include an optional use flag disable-red-menu-bar to keep
SpaceFM from setting the background color of the menu bar to red when
SpaceFM is running with EUID 0.

The file spacefm-1.0.0-disable-red-menu-bar.patch is used for the
patching.

spacefm-1.0.0.ebuild also now requires media-video/ffmpegthumbnailer.