Improvements to the nodejs ebuild
Most of these changes are adopted from iojs, but since nodejs
hasn't evolved the same way as iojs (both build system and src),
a lot of stuff is held back. This is intentional from an ebuild
point of view as well to treat nodejs 0.x as stable builds forward.

This intends to fix the following bugs:
 - https://bugs.gentoo.org/show_bug.cgi?id=542970
 - https://bugs.gentoo.org/show_bug.cgi?id=539342
 - https://bugs.gentoo.org/show_bug.cgi?id=539314
 - https://bugs.gentoo.org/show_bug.cgi?id=540766
 - https://bugs.gentoo.org/show_bug.cgi?id=541656
 - https://bugs.gentoo.org/show_bug.cgi?id=462703
 - https://bugs.gentoo.org/show_bug.cgi?id=540106
 - (pending confirm) https://bugs.gentoo.org/show_bug.cgi?id=533176